+++ evennia/utils/tests/test_utils.py
@@ -99,6 +99,58 @@ class TestMLen(TestCase):
         self.assertEqual(utils.m_len({"hello": True, "Goodbye": False}), 2)
 
 
+class TestDisplayLen(TestCase):
+    """
+    Verifies that display_len behaves like m_len in all situations except those
+    where asian characters are involved.
+    """
+
+    def test_non_mxp_string(self):
+        self.assertEqual(utils.display_len("Test_string"), 11)
+
+    def test_mxp_string(self):
+        self.assertEqual(utils.display_len("|lclook|ltat|le"), 2)
+
+    def test_mxp_ansi_string(self):
+        self.assertEqual(utils.display_len(ANSIString("|lcl|gook|ltat|le|n")), 2)
+
+    def test_non_mxp_ansi_string(self):
+        self.assertEqual(utils.display_len(ANSIString("|gHello|n")), 5)
+
+    def test_list(self):
+        self.assertEqual(utils.display_len([None, None]), 2)
+
+    def test_dict(self):
+        self.assertEqual(utils.display_len({"hello": True, "Goodbye": False}), 2)
+
+    def test_east_asian(self):
+        self.assertEqual(utils.display_len("서서서"), 6)
+
+
+class TestANSIString(TestCase):
+    """
+    Verifies that ANSIString's string-API works as intended.
+    """
+
+    def setUp(self):
+        self.example_raw = "|relectric |cboogaloo|n"
+        self.example_ansi = ANSIString(self.example_raw)
+        self.example_str = "electric boogaloo"
+        self.example_output = "\x1b[1m\x1b[31melectric \x1b[1m\x1b[36mboogaloo\x1b[0m"
+
+    def test_length(self):
+        self.assertEqual(len(self.example_ansi), 17)
+
+    def test_clean(self):
+        self.assertEqual(self.example_ansi.clean(), self.example_str)
+
+    def test_raw(self):
+        self.assertEqual(self.example_ansi.raw(), self.example_output)
+
+    def test_format(self):
+        self.assertEqual(f"{self.example_ansi:0<20}", self.example_output + "000")
+
+
 class TestTimeformat(TestCase):
     """
     Default function header from utils.py: